refactor onEditorStartup
