+++ src/actions/index.ts
@@ -1,2 +1,3 @@
+export { default as onStartup } from './onStartup'
 export { default as onErrorPage } from './onErrorPage'
 export { default as onTestPass } from './onTestPass'
